+++ src/sparkx/JetAnalysis.py
@@ -1,3 +1,12 @@
+#===================================================
+#
+#    Copyright (c) 2023-2024
+#      SPARKX Team
+#
+#    GNU General Public License (GPLv3 or later)
+#
+#===================================================
+    
 import numpy as np
 import fastjet as fj
 import csv
@@ -23,7 +32,7 @@ class JetAnalysis:
     jet_R_: float
         Jet radius parameter.
     jet_eta_range_: tuple
-        Mimimum and maximum pseudorapidity for jet selection.
+        Minimum and maximum pseudorapidity for jet selection.
     jet_pt_range_: tuple
         Minimum transverse momentum for jet selection and maximum transverse
         momentum to write out the jet.
@@ -402,7 +411,7 @@ class JetAnalysis:
                 # create a new file for the first event in the dataset
                 new_file = True
 
-            # perform the jet finiding algorithm
+            # perform the jet finding algorithm
             cluster = fj.ClusterSequence(event_PseudoJets, jet_definition)
             jets = fj.sorted_by_pt(cluster.inclusive_jets(self.jet_pt_range_[0]))
             jets = jet_selector(jets)