+++ src/sparkx/Particle.py
@@ -1,3 +1,12 @@
+#===================================================
+#
+#    Copyright (c) 2023-2024
+#      SPARKX Team
+#
+#    GNU General Public License (GPLv3 or later)
+#
+#===================================================
+    
 import numpy as np
 import math
 from particle import PDGID
@@ -507,10 +516,9 @@ class Particle:
         -------
         pdg : int
         """
-        if np.isnan(self.data_[9]):
+        if (np.isnan(self.data_[9])):
             return np.nan
-        else:
-            return int(self.data_[9])
+        return int(self.data_[9])
 
     @pdg.setter
     def pdg(self,value):
@@ -531,10 +539,9 @@ class Particle:
         -------
         ID : int
         """
-        if np.isnan(self.data_[11]):
+        if (np.isnan(self.data_[11])):
             return np.nan
-        else:
-            return int(self.data_[11])
+        return int(self.data_[11])
 
     @ID.setter
     def ID(self,value):
@@ -548,10 +555,9 @@ class Particle:
         -------
         charge : int
         """
-        if np.isnan(self.data_[12]):
+        if (np.isnan(self.data_[12])):
             return np.nan
-        else:
-            return int(self.data_[12])
+        return int(self.data_[12])
 
     @charge.setter
     def charge(self,value):
@@ -565,10 +571,9 @@ class Particle:
         -------
         ncoll : int
         """
-        if np.isnan(self.data_[13]):
+        if (np.isnan(self.data_[13])):
             return np.nan
-        else:
-            return int(self.data_[13]) 
+        return int(self.data_[13]) 
 
     @ncoll.setter
     def ncoll(self,value):
@@ -610,10 +615,9 @@ class Particle:
         -------
         proc_id_origin : int
         """
-        if np.isnan(self.data_[16]):
+        if (np.isnan(self.data_[16])):
             return np.nan
-        else:
-            return int(self.data_[16])
+        return int(self.data_[16])
 
     @proc_id_origin.setter
     def proc_id_origin(self,value):
@@ -627,10 +631,9 @@ class Particle:
         -------
         proc_type_origin : int
         """
-        if np.isnan(self.data_[17]):
+        if (np.isnan(self.data_[17])):
             return np.nan
-        else:
-            return int(self.data_[17])
+        return int(self.data_[17])
 
     @proc_type_origin.setter
     def proc_type_origin(self,value):
@@ -658,10 +661,9 @@ class Particle:
         -------
         pdg_mother1 : int
         """
-        if np.isnan(self.data_[19]):
+        if (np.isnan(self.data_[19])):
             return np.nan
-        else:
-            return int(self.data_[19])
+        return int(self.data_[19])
 
     @pdg_mother1.setter
     def pdg_mother1(self,value):
@@ -675,10 +677,9 @@ class Particle:
         -------
         pdg_mother2 : int
         """
-        if np.isnan(self.data_[20]):
+        if (np.isnan(self.data_[20])):
             return np.nan
-        else:
-            return int(self.data_[20])
+        return int(self.data_[20])
 
     @pdg_mother2.setter
     def pdg_mother2(self,value):
@@ -694,10 +695,9 @@ class Particle:
         -------
         status : int
         """
-        if np.isnan(self.data_[21]):
+        if (np.isnan(self.data_[21])):
             return np.nan
-        else:
-            return int(self.data_[21])
+        return int(self.data_[21])
 
     @status.setter
     def status(self,value):