+++ src/sparkx/flow/__init__.py
@@ -1,3 +1,12 @@
+#===================================================
+#
+#    Copyright (c) 2023-2024
+#      SPARKX Team
+#
+#    GNU General Public License (GPLv3 or later)
+#
+#===================================================
+    
 from sparkx.flow.ReactionPlaneFlow import ReactionPlaneFlow
 from sparkx.flow.EventPlaneFlow import EventPlaneFlow
 from sparkx.flow.ScalarProductFlow import ScalarProductFlow